Updated Latex fetch request
The previous implementation was beyond hacky and just appended the latex code as a string to the URL. Now we're using an object with proper URL params in requests so that everything is encoded properly. There's still some bugs in loading an SVG with a `+` in blender, though that is likely not to do with this plugin. Perhaps another latex provider should be investigated.